Add examples and bump version (#8)
This pull request updates the `default_avatar_generator` gem with a new
feature, documentation improvements, and a version bump. The main
changes include the addition of example images in the README, updates to
test results, and a version increment to reflect the new changes.

### Documentation Improvements:
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R8-R11):
Added an "Examples" section with a sample image showcasing avatars
generated by the library.

### Test Updates:
*
[`.rspec_status`](diffhunk://#diff-10fae3313daf1aa85d29b18688ebbd7f93b819e9d07072c7a42f716068d570b6L3-R23):
Updated test results to include new test cases for `colors_spec.rb` and
adjusted run times for existing tests.

### Version Update:
*
[`lib/default_avatar_generator/version.rb`](diffhunk://#diff-ba69e3a4eaee31efd2bfd1df41f23d9aeb08d143321d2b8e4b12cf067a4af74fL4-R4):
Updated the version from `0.1.2` to `0.2.0` to indicate the addition of
new features.